backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
pydantic==2.6.1
python-multipart==0.0.9
supabase==2.3.5
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.0.2
pytest-asyncio==0.23.6 